+++ kong/request_termination_attack_test.go
@@ -1,102 +1,18 @@
 // SPDX-License-Identifier: MIT
 // SPDX-FileCopyrightText: 2022 Steadybit GmbH
 
-package services
+package kong
 
 import (
 	"context"
 	"encoding/json"
-	"github.com/kong/go-kong/kong"
 	"github.com/steadybit/attack-kit/go/attack_kit_api"
-	"github.com/steadybit/discovery-kit/go/discovery_kit_api"
 	"github.com/steadybit/extension-kong/config"
-	"github.com/steadybit/extension-kong/routes"
-	"github.com/steadybit/extension-kong/steadybit"
 	"github.com/steadybit/extension-kong/utils"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"testing"
 )
-
-func TestKongServices(t *testing.T) {
-	WithTestContainers(t, []WithTestContainersCase{
-		{
-			Name: "Discover a single service",
-			Test: testDiscoverServices,
-		},
-		{
-			Name: "Discover a single route",
-			Test: testDiscoverRoutes,
-		},
-		{
-			Name: "Kong has no services by default",
-			Test: testDiscoverNoServicesWhenNoneAreConfigured,
-		},
-		{
-			Name: "prepare fails on missing service",
-			Test: testPrepareFailsWhenServiceIsMissing,
-		}, {
-			Name: "prepare fails on unknown instance",
-			Test: testPrepareFailsWhenInstanceIsUnknown,
-		}, {
-			Name: "prepare does not panic on broken JSON",
-			Test: testPrepareNoPanicOnBrokenJson,
-		}, {
-			Name: "prepare configures disabled plugin",
-			Test: testPrepareConfiguresDisabledPlugin,
-		}, {
-			Name: "prepare fails on unknown consumer",
-			Test: testPrepareFailsOnUnknownConsumer,
-		}, {
-			Name: "prepare with a known consumer",
-			Test: testPrepareWithConsumer,
-		}, {
-			Name: "prepare with a route",
-			Test: testPrepareWithRoute,
-		}, {
-			Name: "start enables plugins",
-			Test: testStartEnablesPlugin,
-		}, {
-			Name: "stop deletes plugins",
-			Test: testStopDeletesPlugin,
-		},
-	})
-}
-
-func testDiscoverServices(t *testing.T, instance *config.Instance) {
-	// Given
-	configureService(t, instance, getTestService())
-
-	// When
-	targets := GetServiceTargets(instance)
-
-	// Then
-	assert.NotEmpty(t, targets)
-	target := targets[0]
-	assert.Equal(t, "mockbin", target.Label)
-	assert.Equal(t, []string{"https://mockbin.org:443/request"}, target.Attributes["kong.service.url"])
-	assert.Equal(t, []string{"true"}, target.Attributes["kong.service.enabled"])
-}
-
-func testDiscoverRoutes(t *testing.T, instance *config.Instance) {
-	// Given
-	service := configureService(t, instance, getTestService())
-	configureRoute(t, instance, getTestRoute(service))
-
-	// When
-	targets := routes.GetRouteTargets(instance)
-
-	// Then
-	assert.NotEmpty(t, targets)
-	target := targets[0]
-	assert.Equal(t, "test", target.Label)
-	assert.Equal(t, []string{"/products"}, target.Attributes["kong.route.path"])
-}
-
-func testDiscoverNoServicesWhenNoneAreConfigured(t *testing.T, instance *config.Instance) {
-	targets := GetServiceTargets(instance)
-	assert.Empty(t, targets)
-}
 
 func testPrepareFailsWhenServiceIsMissing(t *testing.T, instance *config.Instance) {
 	// Given
@@ -112,7 +28,7 @@ func testPrepareFailsWhenServiceIsMissing(t *testing.T, instance *config.Instanc
 	require.NoError(t, err)
 
 	// When
-	state, attackKitError := steadybit.PrepareRequestTermination(requestBodyJson)
+	state, attackKitError := PrepareRequestTermination(requestBodyJson)
 	assert.Nil(t, state)
 	assert.Contains(t, attackKitError.Title, "Failed to find service")
 }
@@ -131,14 +47,14 @@ func testPrepareFailsWhenInstanceIsUnknown(t *testing.T, _ *config.Instance) {
 	require.NoError(t, err)
 
 	// When
-	state, attackKitError := steadybit.PrepareRequestTermination(requestBodyJson)
+	state, attackKitError := PrepareRequestTermination(requestBodyJson)
 	assert.Nil(t, state)
 	assert.Contains(t, attackKitError.Title, "Failed to find a configured instance named")
 }
 
 func testPrepareNoPanicOnBrokenJson(t *testing.T, _ *config.Instance) {
 	// When
-	state, attackKitError := steadybit.PrepareRequestTermination([]byte{})
+	state, attackKitError := PrepareRequestTermination([]byte{})
 	assert.Nil(t, state)
 	assert.Contains(t, attackKitError.Title, "Failed to parse request body")
 }
@@ -165,7 +81,7 @@ func testPrepareConfiguresDisabledPlugin(t *testing.T, instance *config.Instance
 	require.NoError(t, err)
 
 	// When
-	state, attackKitError := steadybit.PrepareRequestTermination(requestBodyJson)
+	state, attackKitError := PrepareRequestTermination(requestBodyJson)
 
 	// Then
 	assert.Nil(t, attackKitError)
@@ -203,7 +119,7 @@ func testPrepareFailsOnUnknownConsumer(t *testing.T, instance *config.Instance) 
 	require.NoError(t, err)
 
 	// When
-	state, attackKitError := steadybit.PrepareRequestTermination(requestBodyJson)
+	state, attackKitError := PrepareRequestTermination(requestBodyJson)
 
 	// Then
 	assert.Nil(t, state)
@@ -238,7 +154,7 @@ func testPrepareWithConsumer(t *testing.T, instance *config.Instance) {
 	require.NoError(t, err)
 
 	// When
-	state, attackKitError := steadybit.PrepareRequestTermination(requestBodyJson)
+	state, attackKitError := PrepareRequestTermination(requestBodyJson)
 
 	// Then
 	assert.Nil(t, attackKitError)
@@ -278,7 +194,7 @@ func testPrepareWithRoute(t *testing.T, instance *config.Instance) {
 	require.NoError(t, err)
 
 	// When
-	state, attackKitError := steadybit.PrepareRequestTermination(requestBodyJson)
+	state, attackKitError := PrepareRequestTermination(requestBodyJson)
 
 	// Then
 	assert.Nil(t, attackKitError)
@@ -305,7 +221,7 @@ func testStartEnablesPlugin(t *testing.T, instance *config.Instance) {
 	require.NoError(t, err)
 
 	// When
-	newState, attackKitError := steadybit.StartRequestTermination(startRequestBodyJson)
+	newState, attackKitError := StartRequestTermination(startRequestBodyJson)
 
 	// Then
 	assert.Nil(t, attackKitError)
@@ -319,7 +235,7 @@ func testStartEnablesPlugin(t *testing.T, instance *config.Instance) {
 	assert.Equal(t, "Hello from Kong extension", plugin.Config["message"])
 }
 
-func getSuccessfulPreparationState(t *testing.T, instance *config.Instance) *steadybit.RequestTerminationState {
+func getSuccessfulPreparationState(t *testing.T, instance *config.Instance) *RequestTerminationState {
 	service := configureService(t, instance, getTestService())
 	requestBody := attack_kit_api.PrepareAttackRequestBody{
 		Config: map[string]interface{}{
@@ -336,12 +252,12 @@ func getSuccessfulPreparationState(t *testing.T, instance *config.Instance) *ste
 	prepareRequestBodyJson, err := json.Marshal(requestBody)
 	require.NoError(t, err)
 
-	state, attackKitError := steadybit.PrepareRequestTermination(prepareRequestBodyJson)
+	state, attackKitError := PrepareRequestTermination(prepareRequestBodyJson)
 	require.Nil(t, attackKitError)
 	return state
 }
 
-func getSuccessfulStartState(t *testing.T, instance *config.Instance) *steadybit.RequestTerminationState {
+func getSuccessfulStartState(t *testing.T, instance *config.Instance) *RequestTerminationState {
 	prepareState := getSuccessfulPreparationState(t, instance)
 	encodedState, err := utils.EncodeAttackState(prepareState)
 	require.NoError(t, err)
@@ -350,7 +266,7 @@ func getSuccessfulStartState(t *testing.T, instance *config.Instance) *steadybit
 	})
 	require.NoError(t, err)
 
-	startState, attackKitError := steadybit.StartRequestTermination(startRequestBodyJson)
+	startState, attackKitError := StartRequestTermination(startRequestBodyJson)
 	assert.Nil(t, attackKitError)
 	return startState
 }
@@ -369,65 +285,10 @@ func testStopDeletesPlugin(t *testing.T, instance *config.Instance) {
 	require.NoError(t, err)
 
 	// When
-	attackKitError := steadybit.StopRequestTermination(stopRequestBodyJson)
+	attackKitError := StopRequestTermination(stopRequestBodyJson)
 
 	// Then
 	assert.Nil(t, attackKitError)
 	_, err = client.Plugins.Get(context.Background(), &state.PluginIds[0])
 	assert.Error(t, err)
-}
-
-func getTestService() *kong.Service {
-	return &kong.Service{
-		Enabled:  discovery_kit_api.Ptr(true),
-		Host:     discovery_kit_api.Ptr("mockbin.org"),
-		Name:     discovery_kit_api.Ptr("mockbin"),
-		Path:     discovery_kit_api.Ptr("/request"),
-		Port:     discovery_kit_api.Ptr(443),
-		Protocol: discovery_kit_api.Ptr("https"),
-	}
-}
-
-func getTestRoute(service *kong.Service) *kong.Route {
-	return &kong.Route{
-		Name:    discovery_kit_api.Ptr("test"),
-		Service: service,
-		Hosts:   []*string{discovery_kit_api.Ptr("server1")},
-		Paths:   []*string{discovery_kit_api.Ptr("/products")},
-		Tags:    []*string{discovery_kit_api.Ptr("test")},
-		Methods: []*string{utils.String("GET")},
-	}
-}
-
-func configureService(t *testing.T, instance *config.Instance, service *kong.Service) *kong.Service {
-	client, err := instance.GetClient()
-	require.NoError(t, err)
-
-	createdService, err := client.Services.Create(context.Background(), service)
-	require.NoError(t, err)
-	return createdService
-}
-
-func configureRoute(t *testing.T, instance *config.Instance, route *kong.Route) *kong.Route {
-	client, err := instance.GetClient()
-	require.NoError(t, err)
-
-	createdRoute, err := client.Routes.Create(context.Background(), route)
-	require.NoError(t, err)
-	return createdRoute
-}
-
-func getTestConsumer() *kong.Consumer {
-	return &kong.Consumer{
-		Username: discovery_kit_api.Ptr("test-consumer"),
-	}
-}
-
-func configureConsumer(t *testing.T, instance *config.Instance, consumer *kong.Consumer) *kong.Consumer {
-	client, err := instance.GetClient()
-	require.NoError(t, err)
-
-	createdConsumer, err := client.Consumers.Create(context.Background(), consumer)
-	require.NoError(t, err)
-	return createdConsumer
 }